+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Version 1.14a
 
   - Path handling in afl-multicore improved.
   - Uninitialized variable bugs fixed in afl-collect.
+  - afl-minimize now supports parallel afl-tmin invocation.
 
 Version 1.13a